chore: remove returnValue in output display
The return value from execution in QuickJS depends on the type of code.
If it is a module (containing an import or export statement), the
execution returns the module itself.

Otherwise, the value of the last statement is returned as the result.

Seem no elegant way to change this behavior.
Let's simply remove the returnValue from the UI, it's not necessary anyway.